CHAM-314: Fixed windows warnings in tests.

+++ src/vddsc/tests/listener.c
@@ -299,7 +299,7 @@ create_topic_name(const char *prefix, char *name, size_t size)
 {
     /* Get semi random g_topic name. */
     os_procId pid = os_procIdSelf();
-    int tid = abs(os_threadIdToInteger(os_threadIdSelf()));
+    int tid = abs((int)os_threadIdToInteger(os_threadIdSelf()));
     snprintf(name, size, "%s_pid%"PRIprocId"_tid%d", prefix, pid, tid);
     return name;
 }
